+++ src/main/java/com/hangeulbada/domain/workbookset/dto/QuestionRequestDto.java
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.util.Set;
 
 @Data
 @NoArgsConstructor
@@ -15,6 +14,6 @@ import java.util.Set;
 @Schema(description = "QuestionRequestDto Model")
 public class QuestionRequestDto {
     private String content;
-    private double difficulty;
+    private Integer difficulty;
     private String[] tags;
 }
