+++ src/main/java/com/hangeulbada/domain/workbookset/entity/Workbook.java
@@ -20,7 +20,7 @@ public class Workbook{
     private String teacherId;
     private String title;
     private String description;
-    private double difficulty;
+    private Integer difficulty;
     private int questionNum;
     @Builder.Default
     private List<String> questionIds = new ArrayList<>();
